Lib

Description

All implemented algorithms are imported and benched. In the same function there are also loaded data from files for testing.

Documentation

someFunc :: IO ()

Main function. It reads data files for benching and bench alghoritms