MergeSort

Description

Implementation of merge sort alghoritm.

Documentation

mergeSort :: Ord a => [a] -> [a]

Sorts list of ord elements using merge sort algorithm

merge :: Ord a => [a] -> [a] -> [a]

Function takes 2 sorted lists and merges them into one sorted list