Quicksort

Description

Implementation of quicksort algorithm

Documentation

quicksort :: Ord a => [a] -> [a]

As input get list of elements that can be ordered. In output there is list sorted using quicksort alghoritm.

part :: Ord t => [t] -> t -> ([t], [t], [t])

Part list into tuple containing list of greater, equal and smaller elements commparing to pivot