Regression

Description

Implementation of linear regression alghorithm. Computing aproximation linear function ax + b = y for data set.

Documentation

regress

Arguments

:: Fractional a

=> [a]
First argument: list of arguments for some func

-> [a]
Second argument: list of function results for corresponding argumets (first arg)

-> (a, a)
Resault: tuple representing a and b from ax + b = y

Linear Regression of one variable function

cost :: Fractional a => (a, a) -> [a] -> [a] -> a

returns the sum of the cost of the function and a set of values it is a sum of squares of distance between the given function and all of the given elements

getY :: Fractional a => (a, a) -> a -> a

returns a value of function (f(x) = theta_a * x + theta_b) for given x